add scripting contributions for explorer widget
