+++ src/sql/workbench/parts/dashboard/browser/widgets/explorer/explorerWidget.contribution.ts
